// src/modules/species/interfaces/external-species.interface.ts
export interface ExternalSpecies {
  id?: string;
  classification: string;
  designation: string;
  average_height: string;
  average_lifespan: string;
  hair_colors: string;
  skin_colors: string;
  eye_colors: string;
  homeworld: string;
  language: string;
  people: string[];
  created: string;
  edited: string;
  name: string;
  url: string;
}